Adds PersonName3 possible VR for usage in python3
A new VR is proposed to replace PersonNameBase, PersonName, and PersonNameUnicode in python3. This class should also function properly in python 2.x, but for now it's use has been limited to testing in python3.

This class stores the value provided, uses regular expressions to break the name into it's constituents. However, it also provides encode/decode methods to the user for use during reading/writing/and decoding.

NOTE: It is a quasi-immutable object (i.e. methods that alter the state return a new instance)