Fixes Issue #123 by adding tag numbers to error messages
This is a slight modification to the patch provided by videan42.

With this change, exceptions which occur when traversing a dataset will provide information regarding which tag caused the exception to occur, improving debugging. There was a negligible time penalty associated with this functionality.

The appropriate unit tests have been added as well.