uid: changed generation of uid's
The generation of uids was different if a prefix was passed or not.
It could have happened that when not passing a prefix that the same
uid was generated as the parameters used where not unique enought.

The change additonally takes care of two additional things stated in
Dicom PS3.5 2015a 9.1 UID Encoding Rules

no leading zeros
maximum 64 characters