change __repr__ to use double quotes for numerics
I believe the intention of using a single quoted string as a
representation is to prevent the introduction of floating point rounding
errors.

This change means that DSFloats (etc) can be correctly encoded with the
default python JSON library. Currently, the value is single-quoted in
the JSON, causing a parse error when trying to run json.loads().

Changing to double quotes should preserve the intended effect whilst
allowing JSON-based marshalling to work.