Fix handling of bit-packed PixelData
This commit aims to fix the functionality originally introduced in #293:

* bits were unpacked from the left, which was incorrect, see discussion in
https://github.com/pydicom/pydicom/pull/293#pullrequestreview-109529983.
Correct unpack order implemented now, whereas first pixel from the left in
the image frame corresponds to the first bit from the right in the first
byte of the packed PixelData
* fixed consistency checks verifying the size of the PixelData
* reordered the code to perform unpacking after consistency checks
* removed the dependency on numpy.unpackbits and the associated exception

The functionality of the code was verified using the DICOM Segmentation image
dataset being developed in this PR (3x23x38, which illustrates a case where
bits of the individual frame are not aligned at the byte boundary, and the total
length of pixel bits is less than the number of bits required to encode the bytes
in PixelData): https://github.com/QIICR/dcmqi/pull/334,
and this issue: https://github.com/QIICR/dcmqi/issues/341. In turn, consistency of
the rendering of that dataset with the implementation was confirmed for independent
implementations (Brainlab and DCMTK dcm2pnm).