Fix JSON representation of ImageOrientationPatient
`json.dumps()` uses the output of `__repr__` to encode objects, which
means that in order for JSON encoding to work, `__repr__` must produce
valid JSON if possible.

The string __repr__ of all dicom values that are represented as a string
use double quotes which results in valid JSON.

The only exception is `multival.MultiValue` which produces invalid JSON
with single quotes.

This PR changes the `__repr__` of `MultiValue` to use double quotes.

We've experienced a similar issue before:
https://github.com/pydicom/pydicom/pull/235

Thanks!